palette editor: moved stuff responsible for loading textures into textures.go
