+++ hswidget/palettegridwidget/widget.go
@@ -1,9 +1,6 @@
 package palettegridwidget
 
 import (
-	image2 "image"
-	"image/color"
-
 	"github.com/ianling/giu"
 
 	"github.com/OpenDiablo2/OpenDiablo2/d2common/d2interface"
@@ -79,24 +76,4 @@ func (p *widget) buildGrid() {
 			grid.Build()
 		}),
 	}.Build()
-}
-
-func (p *widget) loadTexture(i int) {
-	state := p.getState()
-
-	rgb := image2.NewRGBA(image2.Rect(0, 0, cellSize, cellSize))
-
-	for y := 0; y < cellSize; y++ {
-		for x := 0; x < cellSize; x++ {
-			col := p.colors[i]
-
-			// nolint:gomnd // opacity
-			rgb.Set(x, y, color.RGBA{R: col.R(), G: col.G(), B: col.B(), A: 255})
-		}
-	}
-
-	p.textureLoader.CreateTextureFromARGB(rgb, func(texture *giu.Texture) {
-		state.texture[i] = texture
-		giu.Context.SetState(p.getStateID(), state)
-	})
 }
